Add triggering and debug data generation.
A triggering pattern draft using the trigger vi of the stream class is
added to the diagram disable structure.
A optional debug boolean can be passed from the host to generate linear
data instead of AI aquisition. The lines are offset from each other all
with slope 1 so they can be differentiated and used as a check that the
deleaving code is working.