Use a 100 element lookup table FIFO for the bundling DRAM FIFO.
There still are some skipped samples. I think this comes from backpressure
building up as the DRAM performance is inconsistent and slows down as the
DRAM fills. I increase the upstream FIFO sizes to accomadate this back
pressure. I had to change the implementation of the Bundling FIFO from
Flipflop to Look-up Table to allow for larger FIFO sizes. A compilation
with a large fliflop FIFO failed.